*** empty log message ***

+++ src/frontends/uade123/playlist.h
@@ -1,6 +1,9 @@
 #ifndef _UADE123_PLAYLIST_H_
 #define _UADE123_PLAYLIST_H_
 
+#include <stdio.h>
+#include <stdlib.h>
+
 #include "chrarray.h"
 
 struct playlist {
@@ -17,7 +20,7 @@ int playlist_random(struct playlist *pl, int enable);
 void playlist_repeat(struct playlist *pl);
 int playlist_empty(struct playlist *pl);
 int playlist_add(struct playlist *pl, char *name, int recursive);
-int playlist_get_next(char *name, int maxlen, struct playlist *pl);
+int playlist_get_next(char *name, size_t maxlen, struct playlist *pl);
 void playlist_flush(struct playlist *pl);
 
 #endif
